refactor(discord-message-config): split into multiple services
