--- src/features/discord/messages/constants/discord-message-config.ts
@@ -1,19 +1,0 @@
-import { IDiscordMessageConfig } from "../../interfaces/discord-message-config";
-
-export const DISCORD_MESSAGE_CONFIG: IDiscordMessageConfig = {
-  command: {
-    error: {
-      imageColor: 15562905,
-      imageUrl: `https://i.ibb.co/5jZmzSB/icons8-error-512.png`,
-    },
-    prefix: `!`,
-    version: {
-      imageColor: 11912416,
-      imageUrl: `https://i.ibb.co/ph17BqP/icons8-artificial-intelligence-512.png`,
-    },
-  },
-  error: {
-    imageColor: 15562905,
-    imageUrl: `https://i.ibb.co/r7PHJS1/icons8-bug-512.png`,
-  },
-};
