+++ src/features/discord/messages/enums/discord-message-config-value-name.enum.ts
@@ -1,0 +1,8 @@
+export enum DiscordMessageConfigValueNameEnum {
+  COMMAND_ERROR_IMAGE_COLOR = `message command error image color`,
+  COMMAND_ERROR_IMAGE_URL = `message command error image url`,
+  COMMAND_VERSION_IMAGE_COLOR = `message command version image color`,
+  COMMAND_VERSION_IMAGE_URL = `message command version image url`,
+  ERROR_IMAGE_COLOR = `message error image color`,
+  ERROR_IMAGE_URL = `message error image url`,
+}
